// app/types/HeaderData.ts
export interface HeaderData {
    hId: number;
    hStatus: string;
    hDomain: string;
    hProject: string;
    hType: string;
    hCustomer: string;
    hRenew: string;
    hContract: string;
    hIssue: number;
    hExpire: number;
    hContractExpire: number;
    hCost: number;
    hBuyer: string;
    hManager: string;
    hActor: string;
    hPurpose: string;
    hWebType: string;
    hWebServer: string;
    hDomainServerIp: string;
    hMemo: string;
    hIsDeleted: boolean;
}